Refactor enemy file
